tests: Move instancetype func tests to dedicated directory
