+++ tests/BUILD.bazel
@@ -96,7 +96,6 @@ go_test(
         "container_disk_test.go",
         "dryrun_test.go",
         "hyperv_test.go",
-        "instancetype_test.go",
         "kubectl_test.go",
         "mdev_configuration_allocation_test.go",
         "pausing_test.go",
@@ -149,7 +148,6 @@ go_test(
         "//pkg/hooks/v1alpha1:go_default_library",
         "//pkg/hooks/v1alpha2:go_default_library",
         "//pkg/hooks/v1alpha3:go_default_library",
-        "//pkg/instancetype:go_default_library",
         "//pkg/libvmi:go_default_library",
         "//pkg/network/dns:go_default_library",
         "//pkg/pointer:go_default_library",
@@ -175,8 +173,6 @@ go_test(
         "//staging/src/kubevirt.io/api/core:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/api/instancetype:go_default_library",
-        "//staging/src/kubevirt.io/api/instancetype/v1alpha1:go_default_library",
-        "//staging/src/kubevirt.io/api/instancetype/v1alpha2:go_default_library",
         "//staging/src/kubevirt.io/api/instancetype/v1beta1:go_default_library",
         "//staging/src/kubevirt.io/api/pool:go_default_library",
         "//staging/src/kubevirt.io/api/pool/v1alpha1:go_default_library",
@@ -201,6 +197,7 @@ go_test(
         "//tests/guestlog:go_default_library",
         "//tests/hotplug:go_default_library",
         "//tests/infrastructure:go_default_library",
+        "//tests/instancetype:go_default_library",
         "//tests/launchsecurity:go_default_library",
         "//tests/libconfigmap:go_default_library",
         "//tests/libdv:go_default_library",
